src/requirements.txt
asgiref==3.7.2
Django==4.2.3
pip==23.2.1
setuptools==68.0.0
sqlparse==0.4.4
typing-extensions==4.7.1
tzdata==2023.3
wheel==0.41.0
psycopg2==2.9.6
celery==5.3.1
django_celery_beat==2.5.0
django-celery-results==2.5.1
watchfiles==0.19.0
flower==2.0.0
django_redis==5.3.0
djangorestframework==3.14.0
markdown==3.4.4
django-filter==23.2
djoser==2.2.0
